Backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
cryptography==41.0.5
# Pillow==10.0.1